Split off AGaugeLabel into its own file.
